// frontend/src/app/interfaces/Usuario.ts
export interface Usuario {
	UsuRut: number;
	UsuDv: string;
	UsuNombres: string;
	UsuAPaterno: string;
	UsuAMaterno: string;
	UsuUsuario: string;
	UsuClave: string;
	CodEstado: number;
	CodTipoUsuario: number;
	FecVigencia: Date;
	UsuEmail: string;
	UsuAvatar: string;
	FecIngReg: any;
	UsuIngReg: string;
	FunIngReg: string;
	FecModReg: any;
	UsuModReg: string;
	FunModReg: string;
}